@@ -12,15 +12,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.4.0",
-    "@angular/cli": "^13.3.11",
-    "@angular/common": "^13.4.0",
-    "@angular/compiler": "^13.4.0",
-    "@angular/core": "^13.4.0",
-    "@angular/forms": "^13.4.0",
-    "@angular/platform-browser": "^13.4.0",
-    "@angular/platform-browser-dynamic": "^13.4.0",
-    "@angular/router": "^13.4.0",
+    "@angular/animations": "^14.3.0",
+    "@angular/cli": "^14.2.11",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
     "@clr/angular": "^3.1.0",
     "@clr/icons": "^3.1.0",
     "@clr/ui": "^3.1.0",
@@ -32,8 +32,8 @@
     "@webcomponents/webcomponentsjs": "^2.0.0"
   },
   "devDependencies": {
-    "@angular/compiler-cli": "^13.4.0",
-    "@angular/language-service": "^13.4.0",
+    "@angular/compiler-cli": "^14.3.0",
+    "@angular/language-service": "^14.3.0",
     "@types/jasmine": "~3.5.4",
     "@types/jasminewd2": "~2.0.8",
     "@types/node": "~13.7.1",
@@ -49,6 +49,6 @@
     "ts-node": "~8.6.2",
     "tslint": "~6.1.0",
     "typescript": "~4.6.4",
-    "@angular-devkit/build-angular": "~13.3.11"
+    "@angular-devkit/build-angular": "^14.2.11"
   }
 }
